requirements.txt
flask
flask-cors
flask-login
sqlalchemy
python-dotenv
